+++ pkg/idp/provider/azure/errors.go
@@ -12,3 +12,21 @@ var invalidConfigError = &microerror.Error{
 func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
 }
+
+var notFoundError = &microerror.Error{
+	Kind: "notFoundError",
+}
+
+// IsNotFound asserts notFoundError.
+func IsNotFound(err error) bool {
+	return microerror.Cause(err) == notFoundError
+}
+
+var notExistError = &microerror.Error{
+	Kind: "notExistError",
+}
+
+// IsNotExist asserts notExistError.
+func IsNotExist(err error) bool {
+	return microerror.Cause(err) == notExistError
+}