+++ pkg/idp/provider/provider_test.go
@@ -11,6 +11,7 @@ func TestReadCredentials(t *testing.T) {
 		name        string
 		location    string
 		credentials []ProviderCredential
+		expectError error
 	}{
 		{
 			name:     "case 0",
@@ -24,8 +25,12 @@ func TestReadCredentials(t *testing.T) {
 	for i, tc := range testCases {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			credentials, err := ReadCredentials(tc.location)
-			if err != nil {
-				t.Fatal(err)
+			if err != tc.expectError {
+				if err != nil {
+					t.Fatal(err)
+				} else {
+					t.Fatalf("Expected error %v", tc.expectError)
+				}
 			}
 			if !reflect.DeepEqual(credentials, tc.credentials) {
 				t.Fatalf("Expected %v to be equal to %v", credentials, tc.credentials)